create new Formatter class
